+++ gs/gs-rendertarget.cpp
@@ -18,7 +18,7 @@
  */
 
 #include "gs-rendertarget.h"
-#include "exceptions.h"
+#include "plugin/exceptions.h"
 extern "C" {
 	#pragma warning( push )
 	#pragma warning( disable: 4201 )
