+++ gs/gs-texture.h
