+++ gs/gs-vertex.cpp
