+++ gs/gs-vertex.h
