+++ mask/mask-resource-animation.h
