+++ mask/mask-resource-effect.h
@@ -18,7 +18,7 @@
 */
 #pragma once
 #include "mask-resource.h"
-#include "gs-effect.h"
+#include "gs/gs-effect.h"
 
 namespace Mask {
 	namespace Resource {
