+++ mask/mask-resource-emitter.h
