+++ mask/mask-resource-image.h
@@ -16,51 +16,37 @@
 * along with this program; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA
 */
+
 #pragma once
 #include "mask-resource.h"
-#include "gs-vertexbuffer.h"
-#include "gs-indexbuffer.h"
+#include "gs/gs-texture.h"
+#include "mask.h"
 
 namespace Mask {
 	namespace Resource {
-		class Mesh : public IBase {
+		class Image : public IBase {
 		public:
-			Mesh(Mask::MaskData* parent, std::string name, std::string file);
-			Mesh(Mask::MaskData* parent, std::string name, obs_data_t* data);
-			virtual ~Mesh();
+			Image(Mask::MaskData* parent, std::string name, obs_data_t* data);
+			Image(Mask::MaskData* parent, std::string name, std::string filename);
+			virtual ~Image();
 
 			virtual Type GetType() override;
-
+			virtual std::shared_ptr<GS::Texture> GetTexture() { return m_Texture; }
 			virtual void Update(Mask::Part* part, float time) override;
 			virtual void Render(Mask::Part* part) override;
 
-			std::shared_ptr<GS::VertexBuffer> GetVertexBuffer() {
-				return m_VertexBuffer;
+			void SwapTexture(gs_texture* tex, bool mineToDestroy = false) {
+				m_Texture = std::make_shared<GS::Texture>(tex, mineToDestroy);
 			}
 
-			vec4 GetCenter() { return m_center; }
-
-			bool GetScreenExtents(gs_rect* r, int screen_width, int screen_height, float trsZ);
-
-		private:
-			void LoadObj(std::string file);
-
 		protected:
-			std::shared_ptr<GS::VertexBuffer>	m_VertexBuffer;
-			std::shared_ptr<GS::IndexBuffer>	m_IndexBuffer;
-			vec4								m_center;
+			std::shared_ptr<GS::Texture> m_Texture;
 
-			// cached in GetScreenExtents
-			std::shared_ptr<Mask::Part>			m_part;
-
-			// for delayed gs creation
-			std::string				m_tempFile;
-			uint8_t*				m_rawVertices;
-			uint8_t*				m_rawIndices;
-			int						m_numIndices;
-
-			vec3 CalculateTangent(const GS::Vertex& v1,
-				const GS::Vertex& v2, const GS::Vertex& v3);
+			// delayed gs creation
+			int				m_width, m_height;
+			gs_color_format m_fmt;
+			std::string		m_tempFile;
+			std::vector<std::vector<uint8_t>> m_decoded_mips;
 		};
 	}
 }
