+++ mask/mask-resource-morph.cpp
@@ -19,9 +19,9 @@
 
 #include "mask-resource-morph.h"
 #include "mask.h"
-#include "exceptions.h"
-#include "plugin.h"
-#include "utils.h"
+#include "plugin/exceptions.h"
+#include "plugin/plugin.h"
+#include "plugin/utils.h"
 extern "C" {
 	#pragma warning( push )
 	#pragma warning( disable: 4201 )
