+++ mask/mask-resource-morph.h
