+++ mask/mask-resource-sequence.h
