+++ mask/mask-resource-sound.cpp
