+++ mask/mask-resource-sound.h
