+++ mask/mask-resource.h
