+++ mask/mask.cpp
@@ -22,10 +22,10 @@
 #include "mask-resource-material.h"
 #include "mask-resource-animation.h"
 #include "mask-resource-sequence.h"
-#include "exceptions.h"
-#include "strings.h"
-#include "plugin.h"
-#include "utils.h"
+#include "plugin/exceptions.h"
+#include "plugin/strings.h"
+#include "plugin/plugin.h"
+#include "plugin/utils.h"
 #include <queue>
 #include <thread>
 extern "C" {
