+++ plugin/strings.h
